Remove restrictions table
