+++ lib/access_check_filter.js
@@ -1,8 +1,8 @@
-'use strict';
+"use strict";
 
+const URI = require('hyperswitch').URI;
 const mwUtil = require('./mwUtil');
 const P = require('bluebird');
-const URI = require('hyperswitch').URI;
 
 function redirect(content, location, options) {
     const vary = content.headers.vary ? `${content.headers.vary}, origin` : 'origin';
@@ -19,62 +19,75 @@ function redirect(content, location, options) {
 
 module.exports = (hyper, req, next, options, specInfo) => {
     const rp = req.params;
-    const titleParamName = options.title || 'title';
-    const checkURIParts = [rp.domain, 'sys', 'page_revisions', 'restrictions', rp.title];
+    const continueRequest = () => next(hyper, req)
+    .then((res) => {
+        if (res.status === 302
+            && (req.query.redirect === false || mwUtil.isNoCacheRequest(req))) {
+            res.status = 200;
+            delete res.headers.location;
+        }
+        return res;
+    });
+
+    if (mwUtil.isNoCacheRequest(req)) {
+        return continueRequest();
+    }
+
+    const titleParamName = options.title ? options.title : 'title';
+    const checkURIParts = [rp.domain, 'sys', 'page_revisions', 'page', rp[titleParamName]];
     if (rp.revision) {
         checkURIParts.push(`${rp.revision}`);
     }
 
-    return P.join(
-        next(hyper, req),
-        hyper.get({ uri: new URI(checkURIParts) })
-        .catch({ status: 404 }, () => null)
-    )
-    .spread((content, restriction) => {
-        if (restriction || content.headers.location) {
-            const revInfo = mwUtil.parseETag(content.headers.etag);
+    return hyper.get({ uri: new URI(checkURIParts) })
+    .then((res) => {
+        const revision = res.body.items[0];
 
-            if (restriction && restriction.body) {
-                mwUtil.applyAccessChecks(restriction.body, revInfo.rev);
+        if (revision.redirect && req.query.redirect !== false) {
+            const htmlPath = [rp.domain, 'sys', 'parsoid', 'html', rp[titleParamName]];
+            if (rp.revision) {
+                htmlPath.push(`${rp.revision}`);
+                if (rp.tid) {
+                    htmlPath.push(`${rp.tid}`);
+                }
             }
-
-            // Use redirect target from restrictions table or content response.
-            const redirectTarget = restriction && restriction.body
-                && parseInt(revInfo.rev, 10) === parseInt(restriction.body.rev, 10)
-                && restriction.body.redirect
-                || content.headers.location && decodeURIComponent(content.headers.location);
-            if (redirectTarget
-                    && req.query.redirect !== false
-                    && !mwUtil.isNoCacheRequest(req)) {
-                const newParams = Object.assign({}, rp);
-                newParams[titleParamName] = redirectTarget;
-                const location = mwUtil.createRelativeTitleRedirect(specInfo.path,
-                    req, {
+            return hyper.get({
+                uri: new URI(htmlPath),
+            })
+            .then((html) => {
+                const redirectTitle = mwUtil.extractRedirect(html.body)
+                    // Redirect detected while saving new HTML.
+                    || html.headers.location;
+                if (redirectTitle) {
+                    const newParams = Object.assign({}, rp);
+                    newParams[titleParamName] = redirectTitle;
+                    const location = mwUtil.createRelativeTitleRedirect(specInfo.path, req, {
                         newReqParams: newParams,
                         titleParamName,
                         dropPathAfterTitle: true,
                     });
 
-                let contentPromise;
-                if (options.attach_body_to_redirect) {
-                    contentPromise = P.resolve(content);
-                } else {
-                    contentPromise = P.resolve({
-                        headers: {
-                            etag: content.headers.etag
+                    let contentPromise;
+                    if (options.attach_body_to_redirect) {
+                        if (specInfo.path.indexOf('html') !== -1) {
+                            contentPromise = P.resolve(html);
+                        } else {
+                            contentPromise = next(hyper, req);
                         }
-                    });
-                }
+                    } else {
+                        contentPromise = P.resolve({
+                            headers: {
+                                etag: html.headers.etag
+                            }
+                        });
+                    }
 
-                return contentPromise.then(content => redirect(content, location, options));
-            } else if (content.status === 302) {
-                // Parsoid.js emits a redirect when the new content is parsed because
-                // there could be a race condition with updating a restriction table.
-                // If the client doesn't want a redirect - remove it.
-                content.status = 200;
-                delete content.headers.location;
-            }
+                    return contentPromise.then(content => redirect(content, location, options));
+                } else {
+                    return next(hyper, req);
+                }
+            });
         }
-        return content;
+        return continueRequest();
     });
 };
