+++ lib/mwUtil.js
@@ -198,47 +198,32 @@ mwUtil.getLimit = (hyper, req) => {
  * if access is restricted, no-op otherwise.
  *
  * Item might be either a revision or a restriction.
- * @param {Object} restrictions the restrictions object or a revision object
- * @param {number} [targetRev] the item revision.
- *                             Used in case item is a restriction to
- *                             indicate the content revision num.
+ * @param {Object} revision the revision object
  */
-mwUtil.applyAccessChecks = (restrictions, targetRev) => {
+mwUtil.applyAccessChecks = (revision) => {
     // Page was deleted - new version of tracking
-    if (restrictions.page_deleted) {
-        if (!targetRev || restrictions.rev <= restrictions.page_deleted) {
-            throw new HTTPError({
-                status: 404,
-                body: {
-                    type: 'not_found#page_revisions',
-                    description: 'Page was deleted'
-                }
-            });
-        }
+    if (revision.page_deleted) {
+        throw new HTTPError({
+            status: 404,
+            body: {
+                type: 'not_found#page_revisions',
+                description: 'Page was deleted'
+            }
+        });
     }
 
-    if (restrictions && Array.isArray(restrictions.restrictions)
-            && restrictions.restrictions.length > 0) {
-        // Page was deleted - old version of tracking
-        if (restrictions.restrictions.indexOf('page_deleted') >= 0) {
-            throw new HTTPError({
-                status: 404,
-                body: {
-                    type: 'not_found#page_revisions',
-                    description: 'Page was deleted'
-                }
-            });
-        }
+    if (revision && Array.isArray(revision.restrictions)
+            && revision.restrictions.length > 0) {
         // Revision restricted
-        if (restrictions.restrictions.indexOf('sha1hidden') >= 0
-            || restrictions.restrictions.indexOf('texthidden') >= 0) {
+        if (revision.restrictions.indexOf('sha1hidden') >= 0
+            || revision.restrictions.indexOf('texthidden') >= 0) {
             throw new HTTPError({
                 status: 403,
                 body: {
                     type: 'access_denied#revision',
                     title: 'Access to resource denied',
-                    description: `Access is restricted for revision ${restrictions.rev}`,
-                    restrictions: restrictions.restrictions
+                    description: `Access is restricted for revision ${revision.rev}`,
+                    restrictions: revision.restrictions
                 }
             });
         }
