+++ sys/page_revisions.yaml
@@ -26,12 +26,3 @@ paths:
       summary: Get the revision metadata for a specific revision
       operationId: getTitleRevision
 
-  /restrictions/{title}{/revision}:
-    get:
-      summary: Check access to the title / revision
-      operationId: getRestrictions
-
-  /restrictions/{title}/{revision}:
-    post:
-      summary: Update restrictions / redirect information for a title / revision.
-      operationId: postRestrictions
