+++ v1/content.yaml
@@ -119,7 +119,7 @@ paths:
 
   /html/{title}:
     x-route-filters:
-      - path: ./lib/revision_table_access_check_filter.js
+      - path: ./lib/access_check_filter.js
         options:
           redirect_cache_control: '{{options.response_cache_control}}'
           attach_body_to_redirect: true
@@ -402,7 +402,7 @@ paths:
 
   /html/{title}/{revision}{/tid}:
     x-route-filters:
-      - path: ./lib/revision_table_access_check_filter.js
+      - path: ./lib/access_check_filter.js
         options:
           redirect_cache_control: '{{options.response_cache_control}}'
           attach_body_to_redirect: true
@@ -536,7 +536,7 @@ paths:
 
   /data-parsoid/{title}/{revision}/{tid}:
     x-route-filters:
-      - path: ./lib/revision_table_access_check_filter.js
+      - path: ./lib/access_check_filter.js
         options:
           redirect_cache_control: '{{options.response_cache_control}}'
           attach_body_to_redirect: true
@@ -638,7 +638,7 @@ paths:
 
   /lint/{title}{/revision}:
     x-route-filters:
-      - path: ./lib/revision_table_access_check_filter.js
+      - path: ./lib/access_check_filter.js
         options:
           redirect_cache_control: '{{options.response_cache_control}}'
       - path: lib/security_response_header_filter.js
