+++ v1/definition.yaml
@@ -14,7 +14,7 @@ info:
 paths:
   /definition/{term}:
     x-route-filters:
-      - path: ./lib/revision_table_access_check_filter.js
+      - path: ./lib/access_check_filter.js
         options:
           title: term
           redirect_cache_control: '{{options.response_cache_control}}'
