+++ v1/pdf.yaml
@@ -14,7 +14,7 @@ info:
 paths:
   /pdf/{title}:
     x-route-filters:
-      - path: ./lib/revision_table_access_check_filter.js
+      - path: ./lib/access_check_filter.js
         options:
           redirect_cache_control: '{{options.cache_control}}'
     get:
