+++ v1/summary.yaml
@@ -14,7 +14,7 @@ info:
 paths:
   /summary/{title}:
     x-route-filters:
-      - path: ./lib/revision_table_access_check_filter.js
+      - path: ./lib/access_check_filter.js
         options:
           redirect_cache_control: '{{options.response_cache_control}}'
     get:
